requirements.txt
absl-py==0.8.1
asgiref==3.2.3
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==1.5.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
Django==3.0
entrypoints==0.3
enum34==1.1.6
gast==0.2.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.7.1
html5lib==0.9999999
image==1.5.27
importlib-metadata==1.2.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==8.0.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.16.1
opencv-python==3.3.0.10
opt-einsum==3.1.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.0
ptyprocess==0.6.0
pycocotools==2.0
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
scikit-learn==0.21.3
scipy==1.3.3
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
flask
subprocess
logging
flask_restplus
Pillow
pytesseract
tarfile
flask-cors
